Split up LVM support into seperate modules
