+++ gems/pending/VolumeManager/LVM.rb
@@ -1,0 +1,8 @@
+require 'VolumeManager/LVM/scanner'
+require 'VolumeManager/LVM/parser'
+require 'VolumeManager/LVM/thin'
+require 'VolumeManager/LVM/volume_group'
+require 'VolumeManager/LVM/physical_volume'
+require 'VolumeManager/LVM/logical_volume'
+require 'VolumeManager/LVM/lv_segment'
+require 'VolumeManager/LVM/lvm2disk'
